+++ native/src/paragraph/TextStyle.cc
@@ -133,17 +133,16 @@ extern "C" JNIEXPORT jobjectArray JNICALL Java_org_jetbrains_skija_paragraph_Tex
     jobjectArray shadowsArr = env->NewObjectArray((jsize) shadows.size(), skija::paragraph::Shadow::cls, nullptr);
     for (int i = 0; i < shadows.size(); ++i) {
         const TextShadow& s = shadows[i];
-        auto shadowObj = env->NewObject(skija::paragraph::Shadow::cls, skija::paragraph::Shadow::ctor, s.fColor, s.fOffset.fX, s.fOffset.fY, s.fBlurRadius);
-        env->SetObjectArrayElement(shadowsArr, i, shadowObj);
-        env->DeleteLocalRef(shadowObj);
+        skija::AutoLocal<jobject> shadowObj(env, env->NewObject(skija::paragraph::Shadow::cls, skija::paragraph::Shadow::ctor, s.fColor, s.fOffset.fX, s.fOffset.fY, s.fBlurSigma));
+        env->SetObjectArrayElement(shadowsArr, i, shadowObj.get());
     }
     return shadowsArr;
 }
 
 extern "C" JNIEXPORT void JNICALL Java_org_jetbrains_skija_paragraph_TextStyle__1nAddShadow
-  (JNIEnv* env, jclass jclass, jlong ptr, jint color, jfloat offsetX, jfloat offsetY, jdouble blurRadius) {
+  (JNIEnv* env, jclass jclass, jlong ptr, jint color, jfloat offsetX, jfloat offsetY, jdouble blurSigma) {
     TextStyle* instance = reinterpret_cast<TextStyle*>(static_cast<uintptr_t>(ptr));
-    instance->addShadow({static_cast<SkColor>(color), {offsetX, offsetY}, blurRadius});
+    instance->addShadow({static_cast<SkColor>(color), {offsetX, offsetY}, blurSigma});
 }
 
 extern "C" JNIEXPORT void JNICALL Java_org_jetbrains_skija_paragraph_TextStyle__1nClearShadows